// Frontend/src/api/generated/model/createActionPlanRequest.ts
/**
 * Generated by orval v7.8.0 🍺
 * Do not edit manually.
 * TMB Control API
 * OpenAPI spec version: v1
 */

export interface CreateActionPlanRequest {
  areaId?: string;
  objectiveId?: string;
  statusId?: number;
  year?: number;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  description?: string | null;
  order?: number;
  startIn?: string;
  finisIn?: string;
}