// Frontend/src/api/generated/model/objectiveActionPlanDto.ts
/**
 * Generated by orval v7.8.0 🍺
 * Do not edit manually.
 * TMB Control API
 * OpenAPI spec version: v1
 */
import type { ObjectiveStatusDto } from './objectiveStatusDto';
import type { ObjectiveAreaDto } from './objectiveAreaDto';
import type { ObjectiveDto } from './objectiveDto';

export interface ObjectiveActionPlanDto {
  id?: string;
  year?: number;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  description?: string | null;
  order?: number;
  startIn?: string;
  finisIn?: string;
  status?: ObjectiveStatusDto;
  area?: ObjectiveAreaDto;
  objective?: ObjectiveDto;
}