// Frontend/src/api/generated/model/objectiveMasterDataDto.ts
/**
 * Generated by orval v7.8.0 🍺
 * Do not edit manually.
 * TMB Control API
 * OpenAPI spec version: v1
 */
import type { ObjectiveYearDto } from './objectiveYearDto';
import type { ObjectiveStatusDto } from './objectiveStatusDto';
import type { ObjectiveAreaDto } from './objectiveAreaDto';
import type { ObjectiveDto } from './objectiveDto';

export interface ObjectiveMasterDataDto {
  /** @nullable */
  years?: ObjectiveYearDto[] | null;
  /** @nullable */
  statuses?: ObjectiveStatusDto[] | null;
  /** @nullable */
  areas?: ObjectiveAreaDto[] | null;
  /** @nullable */
  objectives?: ObjectiveDto[] | null;
}