// Frontend/src/api/generated/model/updateActionPlanRequest.ts
/**
 * Generated by orval v7.8.0 🍺
 * Do not edit manually.
 * TMB Control API
 * OpenAPI spec version: v1
 */

export interface UpdateActionPlanRequest {
  areaId?: string;
  objectiveId?: string;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  description?: string | null;
  year?: number;
  statusId?: number;
  startIn?: string;
  finisIn?: string;
}